Re-arranged LVSUBITEMEDIT structure and changed ListViewSubItemEdit proc to use params for hListivew, dwItem and dwSubItem instead of passing them with the older verison of the LVSUBITEMEDIT structure.
Removed dwWrapAround feild and replaced it with a dwOptions field and added a number of defined constants for those options.
Wrapping options: (previously dwWrapAround) and enhanced that with wrapping horizontally, vertically or both.
Key navigation options: tab keys, arrow keys or both.
Notify options: Sent for navigation: nm_click, nm_dblclk, nm_rclick or nm_rdblclk.

Changed lpControlProc for return value of processed messages: TRUE processes as default with whatever was in wParam at the time, FALSE to reject the message, or >1 to replace wParam with.

Added WM_PASTE as a message sent to user's lpControlProc function.